+++ src/allianceutils/middleware/query_count.py
@@ -1,38 +1,13 @@
 """
-Unfortunately django.db.connection doesn't record queries unless DEBUG=True
-We want this to work in production too so we have to wrap the database connections with our own func
+Django DB instrumentation makes this relatively simple now:
 
-Possible strategies:
-
-- check len(django.db.connection)
-    - this only works if DEBUG is True
-    - Execution flow is:
-        - BaseDatabaseWrapper._prepare_cursor()
-            - BaseDatabaseWrapper.make[_debug]_cursor()
-                - CursorDebugWrapper.execute()
-                - CursorDebugWrapper.executemany()
-
-- Set BaseDatabaseWrapper.force_debug_cursor on each connection to True
-    - see django.test.testcases.TransactionTestCase.assertNumQueries()
-        - see django.test.utils.CaptureQueriesContext
-    - see https://github.com/YPlan/django-perf-rec/blob/2.1.0/django_perf_rec/db.py
-
-The problem with both of these is that they capture the content of every query, not just the number of queries
-This data can get quite large depending on the application
-
-- Set connection.cursor() to be a wrapper function and have this call to the underlying connection.__class__.cursor()
-    - django debug toolbar takes this approach, but doesn't play well with other apps: it doesn't correctly restore an
-      existing connection.cursor attribute
-    - this still works for the cursor though, since unlike connections we never need to restore it to the original state
-
-Instead we dynamically create a new class for the connection to override cursor() and have this call the parent
-
-
-NOTE: We could have used django signals but since this is going to be called on every single SQL query we want to
-avoid the associated overhead
+https://docs.djangoproject.com/en/dev/topics/db/instrumentation/
 """
+from __future__ import annotations
+
 import logging
 from typing import Callable
+from typing import cast
 import warnings
 
 from django.conf import settings
@@ -51,34 +26,51 @@ class QueryCounter:
     def __init__(self):
         self.count = 0
 
-    def __call__(self, execute, sql, params, many, context):
+    def __call__(self, execute: Callable, sql: str, params: dict, many, context):
         self.count += 1
         return execute(sql, params, many, context)
+
+
+class QueryCountHttpRequest(HttpRequest):
+    _querycountmiddleware: QueryCountMiddleware
+    QUERY_COUNT_WARNING_THRESHOLD: int
 
 
 class QueryCountMiddleware:
     get_response: Callable
 
-    def __init__(self, get_response: Callable):
+    def __init__(self, get_response: Callable[[HttpRequest], HttpResponse]):
         self.get_response = get_response
 
     def __call__(self, request: HttpRequest) -> HttpResponse:
-        # check for QueryCountMiddleware being included twice
-        if getattr(request, 'querycountmiddleware', None) not in (None, self):
+        if getattr(request, '_querycountmiddleware', None) not in (None, self):
             msg = "QueryCountMiddleware appears to be already initialised (did you include QueryCountMiddleware multiple times?)"
             warnings.warn(msg, RuntimeWarning)
             return self.get_response(request)
 
-        request.querycountmiddleware = self
-        request.querycountmiddleware_query_count = 0
-        request.QUERY_COUNT_WARNING_THRESHOLD = getattr(settings, 'QUERY_COUNT_WARNING_THRESHOLD', DEFAULT_QUERY_COUNT_WARNING_THRESHOLD)
+        # we patch request to add some query count attributes
+        request_qc = cast(QueryCountHttpRequest, request)
 
+        # set this to avoid multiple QueryCountMiddlewares being included
+        request_qc._querycountmiddleware = self
+        request_qc.QUERY_COUNT_WARNING_THRESHOLD = getattr(
+            settings,
+            "QUERY_COUNT_WARNING_THRESHOLD",
+            DEFAULT_QUERY_COUNT_WARNING_THRESHOLD)
         counter = QueryCounter()
+        # request._query_counter = counter  # we expose this for testing
         with connection.execute_wrapper(counter):
-            response = self.get_response(request)
-        query_count = counter.count
+            response = self.get_response(request_qc)
 
-        if getattr(request, 'QUERY_COUNT_WARNING_THRESHOLD', 0) and query_count >= request.QUERY_COUNT_WARNING_THRESHOLD:
-            logger.warning(f'excessive query count: request "{request.method} {request.path}" ran {query_count} queries')
+        if request_qc.QUERY_COUNT_WARNING_THRESHOLD and counter.count >= request_qc.QUERY_COUNT_WARNING_THRESHOLD:
+            logger.warning(f"excessive query count: request '{request_qc.method} {request_qc.path}' ran {counter.count} queries")
 
         return response
+
+    @classmethod
+    def set_threshold(cls, request: HttpRequest, threshold: int):
+        cast(QueryCountHttpRequest, request).QUERY_COUNT_WARNING_THRESHOLD = threshold
+
+    @classmethod
+    def increase_threshold(cls, request: HttpRequest, increment: int):
+        cast(QueryCountHttpRequest, request).QUERY_COUNT_WARNING_THRESHOLD += increment